Dog's Age
Max is having a dog . he want  to find the age of the dog  with respect to the age of human. he came to know that , the age of the  dog is measured with respect to human  has a formula to proceed.

Example: 1 year of life span of dog is same as seven years of life span of human being, now calculate the age of MAX dog.